+++ GameWorld.hpp
@@ -11,7 +11,7 @@
 
 class GameWorld {
 public:
-	GameWorld(intptr_t address) {
+	GameWorld(uint64_t address) {
 		this->address = address;
 	}
 
@@ -20,7 +20,7 @@ public:
 	std::vector<Player> GetPlayers();
 	std::vector<WorldLootItem> GetLoot();
 
-	intptr_t address;
+	uint64_t address;
 
 private:
 	std::optional<std::vector<Player>> cachedPlayers = std::nullopt;
