+++ IMemoryInterface.h
@@ -17,17 +17,17 @@
 __interface IMemoryInterface
 {
 public:
-    intptr_t GetBaseAddress();
+    uint64_t GetBaseAddress();
     bool UpdateProcessId(const wchar_t* processName);
-    bool ReadRaw(intptr_t address, void* pBuffer, unsigned long size);
-    bool WriteRaw(intptr_t address, void* pBuffer, unsigned long size);
-    intptr_t GetModuleBase();
+    bool ReadRaw(uint64_t address, void* pBuffer, unsigned long size);
+    bool WriteRaw(uint64_t address, void* pBuffer, unsigned long size);
+    uint64_t GetModuleBase();
     bool SetTargetModule(wchar_t* moduleName);
 };
 
 namespace Memory {
     template <typename T>
-    bool Read(IMemoryInterface* pMemoryInterface, intptr_t address, T* data) {
+    bool Read(IMemoryInterface* pMemoryInterface, uint64_t address, T* data) {
         if (!address) {
             return false;
         }
@@ -36,7 +36,7 @@ namespace Memory {
     }
 
     template <typename T>
-    T ReadValue(IMemoryInterface* pMemoryInterface, intptr_t address) {
+    T ReadValue(IMemoryInterface* pMemoryInterface, uint64_t address) {
         T t{};
         if (!address) {
             return t;
@@ -47,7 +47,7 @@ namespace Memory {
     }
 
     template <typename T>
-    T ReadChain(IMemoryInterface* pMemoryInterface, intptr_t baseAddress, std::vector<intptr_t> chain) {
+    T ReadChain(IMemoryInterface* pMemoryInterface, uint64_t baseAddress, std::vector<uint64_t> chain) {
         T result{};
         size_t chainSize = chain.size();
         if (chainSize <= 1) {
@@ -56,7 +56,7 @@ namespace Memory {
 
         for (size_t i = 0; i < chainSize; i++) {
             if (i != chainSize - 1) {
-                baseAddress = Memory::ReadValue<intptr_t>(pMemoryInterface, baseAddress + chain[i]);
+                baseAddress = Memory::ReadValue<uint64_t>(pMemoryInterface, baseAddress + chain[i]);
             }
             else {
                 pMemoryInterface->ReadRaw(baseAddress + chain[i], &result, sizeof(T));
@@ -67,22 +67,22 @@ namespace Memory {
     }
 
     template <typename T>
-    bool Write(IMemoryInterface* pMemoryInterface, intptr_t address, const T& data) {
+    bool Write(IMemoryInterface* pMemoryInterface, uint64_t address, const T& data) {
         return pMemoryInterface->WriteRaw(address, (void*)&data, sizeof(T));
     }
 
     template <typename T>
-    std::tuple<int, T*> ReadList(IMemoryInterface* pMemoryInterface, intptr_t address) {
+    std::tuple<int, T*> ReadList(IMemoryInterface* pMemoryInterface, uint64_t address) {
         int length = Memory::ReadValue<int>(pMemoryInterface, address + 0x18);
         if (length == 0) {
             return std::tuple<int, T*>{0, nullptr};
         }
 
         T* buffer = new T[length];
-        pMemoryInterface->ReadRaw(Memory::ReadValue<intptr_t>(pMemoryInterface, address + 0x10) + 0x20, buffer, sizeof(T) * length);
+        pMemoryInterface->ReadRaw(Memory::ReadValue<uint64_t>(pMemoryInterface, address + 0x10) + 0x20, buffer, sizeof(T) * length);
 
         return std::tuple<int, T*>{length, buffer};
     }
 
-    std::wstring ReadString(IMemoryInterface* pMemoryInterface, intptr_t address, bool sanitize = true);
+    std::wstring ReadString(IMemoryInterface* pMemoryInterface, uint64_t address, bool sanitize = true);
 }
