+++ Player.hpp
@@ -13,10 +13,10 @@
 
 class Player {
 public:
-	Player(intptr_t address);
+	Player(uint64_t address);
 
 	ProfileInfo GetProfileInfo();
-	intptr_t GetSkeletonTransformListValues();
+	uint64_t GetSkeletonTransformListValues();
 	Vector3 GetPosition();
 	Vector3 GetBone(EBone bone);
 	void DrawBones(float alpha, ProfileInfo& localPlayerInfo);
@@ -26,13 +26,13 @@ public:
 	float distance;
 	float distance2d;
 	bool isInImportantRange;
-	intptr_t address;
+	uint64_t address;
 
 private:
-	std::map<EBone, intptr_t> cachedBones;
+	std::map<EBone, uint64_t> cachedBones;
 	std::optional<ProfileInfo> cachedProfileInfo = std::nullopt;
-	intptr_t cachedEFTPlayerClassAddress;
-	intptr_t cachedTransformAddress;
+	uint64_t cachedEFTPlayerClassAddress;
+	uint64_t cachedTransformAddress;
 	std::optional<InventoryController> cachedInventoryController = std::nullopt;
 	std::optional<Color> cachedColor = std::nullopt;
