+++ WorldLootItem.cpp
@@ -4,7 +4,7 @@
 
 bool WorldLootItem::IsContainer() {
 	if (!this->cachedIsContainer.has_value()) {
-		this->cachedIsContainer = Memory::ReadValue<intptr_t>(Global::pMemoryInterface, this->address + 0x18) == 0;
+		this->cachedIsContainer = Memory::ReadValue<uint64_t>(Global::pMemoryInterface, this->address + 0x18) == 0;
 	}
 
 	return this->cachedIsContainer.value();
@@ -12,7 +12,7 @@ bool WorldLootItem::IsContainer() {
 
 Vector3 WorldLootItem::GetPosition() {
 	if (!this->cachedTransformAddress) {
-		this->cachedTransformAddress = Memory::ReadValue<intptr_t>(Global::pMemoryInterface, this->address + 0x10);
+		this->cachedTransformAddress = Memory::ReadValue<uint64_t>(Global::pMemoryInterface, this->address + 0x10);
 	}
 
 	return Memory::ReadValue<Vector3>(Global::pMemoryInterface, this->cachedTransformAddress + 0xA90);
@@ -20,7 +20,7 @@ Vector3 WorldLootItem::GetPosition() {
 
 std::optional<std::wstring> WorldLootItem::GetId() {
 	if (!this->cachedId.has_value()) {
-		std::wstring readString = Memory::ReadString(Global::pMemoryInterface, Memory::ReadValue<intptr_t>(Global::pMemoryInterface, this->address + 0x68));
+		std::wstring readString = Memory::ReadString(Global::pMemoryInterface, Memory::ReadValue<uint64_t>(Global::pMemoryInterface, this->address + 0x68));
 		if (IS_VALID_WSTRING(readString)) {
 			this->cachedId = readString;
 		}
